weighted fit with std errors (not working yet)
Implemented the passing through of the statistical errors in the fit data down the fit process.

#sigma=weights

is commented out, as it works worse.

One problem also was zeroes in the errors at smaller energies. By those would be divided for chi-sqr. A function was implemented to assign the minimum of the errors that is bigger than zero, to all those that are zero.

Might be interesting, but needs more work.