Annotations and better statusbar implementation, minor design changes
added functionality for annotations to be put in the top left corner of the figure. LATEX works, compiling errors of LATEX lead to an error message in the statusbar and stops the annotation to be shown any further, until LATEX error is fixed.

Other potential errors, when refreshing the figure have been put in a try-except-block with an adjusted error message in for the status bar.

Minor design changes, sizing.