saving data error fixed
shifted data was saved as raw data! fixed here

raw data is being saved and a copy is created to work with. raw data will be saved after selecting the process.

It was intended to work like this from the beginning, but data[:, 0:2] didn not create a copy. data[:, 0:2][:] does create a copy.