minor
weighted fit is now optional
this feature can be saved
saving and loading changed for aefit

NavigationToolbar has been introduced for the figureCanvas. It makes minor alterations nice and easy, like changing color, marker size, etc. And especially zooming is nice. However it is in conflict with prior zooming and the little marker one can set. So for now it is not visible.