v-tabs as meta seperator and adjustments for export
instead of ',' a v-tab ('\v') is now being used as a seperator for metadata. Previously metadata was corrupted if a comma was used in notes or annotation. This is fixed now. This also makes previous files' metadata unreadable.

exporting data now properly prints out  other fits as there was a typecast issue as a type float64 does not pass a if var is float: check. float was changed throuout the program to np.float64, and the dsb now returns also this type when dsb.value() is called. This however did not resolve the issue, so the check was not to wether or not it is float, but to wether or not it is not nan and that it is finite.